Add a button to set calculated coordinates as a target
Inside the app, there are some tools whic give back coordinates. It is a pain to manually copy these coordinates over to the compass. Therfore, a new button has been added to set calculated coordinates as the compass target.

Also a small programming typo which would result in a broken translation for the German language was fixed.