Update the review handling
When asking for a review, instead of only giving yes and no as the options, it is now also possible to choose for later. This wil reset the usage counter so a review will be asked again at a later time.

Also, tapping outside of the alert will no longer close it.

Refs: #19