Update reviewing
When clicking the menu item, instead of asking for an in-app review, the app will open the review page in the Google Play Store. The old way, opening the in-app review, would not respond if a review was already posted.

Also, the Store Review Plugin was updated to fix issues with Android 11.